requirements.txt
pygame==2.5.2
numpy==1.26.2
openai==1.5.0
python-dotenv==1.0.0
opencv-python==4.11.0.86
